requirements.txt
Flask
Flask-Bcrypt
flask-cors==5.0.0
Flask-Login
Flask-Mail
Flask-Migrate
Flask-MySQLdb
Flask-SQLAlchemy
Flask-WTF
Jinja2
mysql-connector-python
PyJWT
python-dotenv
pytz
Werkzeug
APScheduler
pandas
PyMySQL
pytest
openpyxl

# #to run this is the command
# pip install -r requirements.txt

